implement hidden states
This turned out to be easier than anticipated, I think. Since each 'block' produces the hidden states, it is easy to grab them.

That is what wav2vec and the huggingface transformers do in general. It is a bit obscure but it follows the logic here:
https://github.com/huggingface/transformers/blob/main/src/transformers/models/wav2vec2/modeling_wav2vec2.py

and in timm: https://github.com/huggingface/pytorch-image-models/blob/960202cfcc5630374908c97154ddba4d75c3356b/timm/models/deit.py#L76

I didn't do extensive testing, but I did check that it runs and that the shapes are all ok.

What I did not do was implement the option for embeddings to also be the (learnt/optimized) weighted sum